# api/docs/swagger.yaml
definitions:
...
  handler.CreatePostRequest:
    properties:
      body:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - body
    - slug
    - status
    - title
    type: object
...
  handler.UpdatePostRequest:
    properties:
      body:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - body
    - slug
    - status
    - title
    type: object
...
  view.PostResponse:
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      published_at:
        type: string
      slug:
        type: string
      status:
        type: string
      summary:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/posts:
    get:
      description: List all published blog posts with optional pagination and tag
        filtering
      parameters:
      - description: 'Number of posts per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: 'Number of posts to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Filter posts by tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of published posts
          schema:
            items:
              $ref: '#/definitions/view.PostResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all published blog posts
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Create a new blog post (admin only)
      parameters:
      - description: Post request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Post created successfully
          schema:
            $ref: '#/definitions/view.PostResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new blog post
      tags:
      - Posts
  /api/posts/{id}:
    delete:
      description: Delete a blog post by ID (admin only)
      parameters:
      - description: Post ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Post deleted successfully
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a blog post
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Update a blog post by ID (admin only)
      parameters:
      - description: Post ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            $ref: '#/definitions/view.PostResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - admin role required
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a blog post
      tags:
      - Posts
  /api/posts/{slug}:
    get:
      description: Get a blog post by slug (public posts visible to all, drafts only
        to admin)
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post found
          schema:
            $ref: '#/definitions/view.PostResponse'
        "400":
          description: Missing slug parameter
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Post not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a blog post by slug
      tags:
      - Posts
...
swagger: "2.0"